Revert bad commits
