+++ docs/source/index.rst
@@ -1,49 +1,43 @@
+.. _index:
+
+.. iexfinance documentation master file, created by
+   sphinx-quickstart on Sat Jan 27 14:14:10 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+.. include:: ../../README.rst
+
+.. _index.documentation:
+
+
 Welcome to iexfinance's documentation!
 ======================================
 
-``iexfinance``'s documentation is organized into the following sections:
-
-- :ref:`getting_started` - Installation instructions, basic usage information,
-  migrating to IEX Cloud
-- :ref:`endpoints` - 1:1 mirror of the `IEX Cloud documentation <https://iexcloud.io/api/>`__
-- :ref:`package_info` - additional package information, developer / testing
-  documentation
-
-.. note:: ``iexfinance`` version 0.4.0 introduces a refactored :ref:`refdata`
-          and :ref:`iexdata` (formerly IEX Stats and IEX Market Data). For more
-          information, see :ref:`migrating`.
-
-Full Contents
--------------
-
-.. _getting_started:
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Getting Started
-
-   whatsnew
-   migrating
-   install
-   configuration
-   usage
-
-.. _endpoints:
+Contents
+--------
 
 .. toctree::
    :maxdepth: 2
+   :caption: Getting Started
+
+   whatsnew
+   install
+   usage
+
+.. toctree::
+   :maxdepth: 3
    :caption: Endpoints
 
    account
    stocks
-   refdata
-   iexdata
+   historical
+   ref
+   market
+   stats
    apistatus
 
-.. _package_info:
-
 .. toctree::
-   :maxdepth: 1
+   :maxdepth: 2
    :caption: Package Information
 
    about
