+++ docs/source/migrating.rst
@@ -2,16 +2,3 @@
 
 Migrating to IEX Cloud
 ======================
-
-
-
-
-.. _migrating.ref_data:
-
-Reference Data
---------------
-
-- The ``ref-data/symbols`` endpoint has changed in IEX Cloud. In the v1
-  (legacy) Developer API, this endpoint returned the list of symbols IEX
-  supports for trading. In IEX cloud, it returns the list of symbols IEX
-  supports for *api calls*.
