+++ docs/source/whatsnew.rst
@@ -8,6 +8,8 @@ New features, bug fixes, and improvements for each release.
 
 .. include:: whatsnew/v0.4.0.txt
 
+.. include:: whatsnew/v0.3.6.txt
+
 .. include:: whatsnew/v0.3.5.txt
 
 .. include:: whatsnew/v0.3.4.txt
