+++ docs/source/whatsnew/v0.3.6.txt
@@ -1,0 +1,16 @@
+.. _whatsnew_036:
+
+
+v0.3.6 (TBD)
+-----------------------
+
+This is a minor release from 0.3.5.
+
+
+Bug Fixes
+~~~~~~~~~
+
+- Make pandas indexing consistent for ``get_historical_intraday``, using
+  ``pandas.DatetimeIndex`` instead of ``str`` (GH95_)
+
+.. _GH95: https://github.com/addisonlynch/iexfinance/issues/95
