+++ iexfinance/base.py
@@ -24,9 +24,6 @@ class _IEXBase(object):
         "iexcloud-v1": "https://cloud.iexapis.com/v1/"
     }
 
-    _VALID_FORMATS = ('json', 'pandas')
-    _VALID_CLOUD_VERSIONS = ("iexcloud-beta", "iexcloud-v1")
-
     def __init__(self, **kwargs):
         """ Initialize the class
 
@@ -43,8 +40,7 @@ class _IEXBase(object):
         json_parse_float: datatype, default float, optional
             Desired floating point parsing datatype
         output_format: str, default "json", optional
-            Desired output format (json or pandas DataFrame). This can also be
-            set using the environment variable ``IEX_OUTPUT_FORMAT``.
+            Desired output format (json or pandas DataFrame)
         token: str, optional
             Authentication token (required for use with IEX Cloud)
         """
@@ -53,24 +49,20 @@ class _IEXBase(object):
         self.session = _init_session(kwargs.get("session"))
         self.json_parse_int = kwargs.get("json_parse_int")
         self.json_parse_float = kwargs.get("json_parse_float")
-        self.output_format = kwargs.get("output_format",
-                                        os.getenv("IEX_OUTPUT_FORMAT", 'json'))
-        if self.output_format not in self._VALID_FORMATS:
-            raise ValueError("Please enter a valid output format ('json' "
-                             "or 'pandas').")
-        self.token = kwargs.get("token")
+        self.output_format = kwargs.get("output_format", 'json')
+        self.api_key = kwargs.get("token")
 
         # Get desired API version from environment variables
         # Defaults to v1 API
         self.version = os.getenv("IEX_API_VERSION")
-        if self.version in self._VALID_CLOUD_VERSIONS:
-            if self.token is None:
-                self.token = os.getenv('IEX_TOKEN')
-            if not self.token or not isinstance(self.token, str):
+        if self.version in ("iexcloud-beta", "iexcloud-v1"):
+            if self.api_key is None:
+                self.api_key = os.getenv('IEX_API_KEY')
+            if not self.api_key or not isinstance(self.api_key, str):
                 raise auth_error('The IEX Cloud API key must be provided '
-                                 'either through the token variable or '
+                                 'either through the api_key variable or '
                                  'through the environmental variable '
-                                 'IEX_TOKEN.')
+                                 'IEX_API_KEY.')
         else:
             self.version = 'v1'
 
@@ -137,7 +129,7 @@ class _IEXBase(object):
             If problems arise when making the query
         """
         params = self.params
-        params['token'] = self.token
+        params['token'] = self.api_key
         for i in range(self.retry_count+1):
             response = self.session.get(url=url, params=params)
             if response.status_code == requests.codes.ok:
@@ -187,7 +179,6 @@ class _IEXBase(object):
             A response object
         """
         url = self._prepare_query()
-        print("URL: %s" % url)
         data = self._execute_iex_query(url)
         return self._output_format(data, fmt_j=fmt_j, fmt_p=fmt_p)
