+++ iexfinance/stocks/__init__.py
@@ -1,4 +1,4 @@
-from iexfinance.stocks.base import Stock # noqa
+from iexfinance.stocks.base import StockReader
 from iexfinance.stocks.collections import CollectionsReader
 from iexfinance.stocks.crypto import CryptoReader
 from iexfinance.stocks.historical import HistoricalReader, IntradayReader
@@ -14,9 +14,34 @@ from iexfinance.utils import _sanitize_dates
 # and conditions of use
 
 
+def Stock(symbols=None, **kwargs):
+    """
+    Function to to retrieve data from the IEX Stocks endpoints
+
+    Parameters
+    ----------
+    symbols: str or list
+        A string or list of strings that are valid symbols
+    output_format: str, default 'json', optional
+        Desired output format for requests
+    kwargs:
+        Additional Request Parameters (see base class)
+    Returns
+    -------
+    stock.StockReader
+        A StockReader instance
+    """
+    if isinstance(symbols, str) and symbols:
+        return StockReader([symbols], **kwargs)
+    elif isinstance(symbols, list) and 0 < len(symbols) <= 100:
+        return StockReader(symbols, **kwargs)
+    else:
+        raise ValueError("Please input a symbol or list of symbols")
+
+
 def get_historical_data(symbols, start=None, end=None, **kwargs):
     """
-    Function to obtain historical date for a symbol or list of
+    Top-level function to obtain historical date for a symbol or list of
     symbols. Return an instance of HistoricalReader
 
     Parameters
@@ -41,7 +66,7 @@ def get_historical_data(symbols, start=None, end=None, **kwargs):
 
 def get_historical_intraday(symbol, date=None, **kwargs):
     """
-    Function to obtain intraday one-minute pricing data for one
+    Top-level function to obtain intraday one-minute pricing data for one
     symbol on a given date (defaults to current date)
 
     Parameters
@@ -79,7 +104,7 @@ def get_sector_performance(**kwargs):
 
 def get_collections(collection_name, collection_type="tag", **kwargs):
     """
-    Function for obtaining data from the Collections endpoint
+    Top-level function for obtaining data from the Collections endpoint
 
     Parameters
     ----------
@@ -96,7 +121,7 @@ def get_collections(collection_name, collection_type="tag", **kwargs):
 
 def get_market_volume(**kwargs):
     """
-    Function for obtaining market volume data
+    Top-level function for obtaining market volume data
 
     Data Weighting: 1 per call
 
@@ -108,24 +133,14 @@ def get_market_volume(**kwargs):
 
 def get_crypto_quotes(**kwargs):
     """
-    Function for obtaining all available cryptocurrency quotes
+    Top-level function for obtaining all available cryptocurrency quotes
     """
     return CryptoReader(**kwargs).fetch()
 
 
 def get_todays_earnings(**kwargs):
     """
-    DEPRECATED: renamed get_earnings_today
-    """
-    import warnings
-    warnings.warn("get_today_earnings has been renamed get_earnings_today and "
-                  "will be deprecated in v0.4.1")
-    return EarningsReader(**kwargs).fetch()
-
-
-def get_earnings_today(**kwargs):
-    """
-    Function for obtaining all earnings results which are released on
+    Top-level function for obtaining all earnings results which are released on
     the current date.
     """
     return EarningsReader(**kwargs).fetch()
@@ -133,7 +148,7 @@ def get_earnings_today(**kwargs):
 
 def get_ipo_calendar(period="upcoming-ipos", **kwargs):
     """
-    Function for obtaining today's and upcoming IPOs
+    Top-level function for obtaining today's and upcoming IPOs
 
     Parameters
     ----------
@@ -145,7 +160,7 @@ def get_ipo_calendar(period="upcoming-ipos", **kwargs):
 
 def get_market_gainers(**kwargs):
     """
-    Function for obtaining top 10 market gainers from the
+    Top-level function for obtaining top 10 market gainers from the
     Stocks list endpoint
     """
     return MoversReader(mover='gainers', **kwargs).fetch()
@@ -153,7 +168,7 @@ def get_market_gainers(**kwargs):
 
 def get_market_losers(**kwargs):
     """
-    Function for obtaining top 10 market losers from the
+    Top-level function for obtaining top 10 market losers from the
     Stocks list endpoint
     """
     return MoversReader(mover='losers', **kwargs).fetch()
@@ -161,7 +176,7 @@ def get_market_losers(**kwargs):
 
 def get_market_most_active(**kwargs):
     """
-    Function for obtaining top 10 most active symbols from
+    Top-level function for obtaining top 10 most active symbols from
     the Stocks list endpoint
     """
     return MoversReader(mover='mostactive', **kwargs).fetch()
@@ -169,7 +184,7 @@ def get_market_most_active(**kwargs):
 
 def get_market_iex_volume(**kwargs):
     """
-    Function for obtaining the 10 symbols with the highest
+    Top-level function for obtaining the 10 symbols with the highest
     IEX volume from the Stocks list endpoint
     """
     return MoversReader(mover='iexvolume', **kwargs).fetch()
@@ -177,7 +192,7 @@ def get_market_iex_volume(**kwargs):
 
 def get_market_iex_percent(**kwargs):
     """
-    Function for obtaining the 10 symbols with the highest
+    Top-level function for obtaining the 10 symbols with the highest
     percent change on the IEX exchange from the Stocks list endpoint
     """
     return MoversReader(mover='iexpercent', **kwargs).fetch()
@@ -185,7 +200,7 @@ def get_market_iex_percent(**kwargs):
 
 def get_market_in_focus(**kwargs):
     """
-    Function for obtaining top 10 in focus symbols from the
+    Top-level function for obtaining top 10 in focus symbols from the
     Stocks list endpoint
     """
     return MoversReader(mover='infocus', **kwargs).fetch()
