--- pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-markers=
-    cloud: mark test for use only when IEX Cloud is the API version
-    legacy: mark test for use only when legacy 1.0 API is the API version
